api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
stripe
passlib[bcrypt]
pyjwt